requirements.txt
Django==3.0.6
djangorestframework
psycopg2
Pillow
flake8
django-mptt